part one of the new 'bulk' transfer code.  What's left is to write the implementation for the piped soundserver.
I'll probably scrap the old song xfer code completely, as it's more
complex than I think we really need.

All of the new code here is not currently in the executtion path.  When
the _null server is up to speed,  we should be at the point where the
main soundserver loop is extractable and portable.  :)

 - Pizza