* Fixed line drawing bug in pic0 drawing
This bug was responsible for some heap corruption, but not for all of it.
If we do two memory checks simultaneously (with dmalloc and an SDL thread),
dmalloc still bails out; this happens as soon as the second thread is started
(a minimal test case is at http://signine.yi.org/sdlmem.tar.gz). Any ideas?

Regarding #187: There appears to be one actual difference between the
dynviews we're supposed to collide with and those we're not: The solid ones
are present during the first Animate() after a DrawPic(). There must be
something more to this, though, as this call to Animate() also draws egoObj
in some cases.

-- Christoph