* Added Nikita's patch to improve broken resource detection in decrypt1() * (a) GGI driver now works on arbitrary resolutions * Added --without-sdl and --without-ggi options * Added sci_sched_yield() function * (b) Added sci_malloc(), sci_realloc(), sci_calloc(), sci_free() functions
(a) will not scale the picture, but, rather, center it. It's relevant for
fb consoles which can't run arbitrary resolutions.

(b) hasn't been tested yet, but it should suspend execution if insufficient
memory is available and try again each second, making sure the function
either does not terminate or terminates successfully.


-- Christoph