* Fixed Win32 specific code in main.c * (a) Added #defineable update debug code to operations.c (more to come) * (b) Fixed soundserver_sdl to lock the mutex before waiting on it
(a) indicates that it's basically going fine- multiple draws are being
aggregated into one big rectangle- but that some stuff appears to be dirtified
which shouldn't be, effectively causing too big an area to be redrawn.

(b) fixes the error messages (which made it unuseable on HZ 1024 systems),
but causes latencies of one second. The sched_yield() approach would appear
to be better, although still not acceptable in quality.

How about using pipes and select()ing on them? Appears to work great for the
UNIX sound server...
Any other suggestions?

-- Christoph