In src/engine/klists.c:
	Added workaround for a bug in Codename: Iceman at the map table.

In src/engine/vm.c:

	(a) Disabled check for number of allocated hunk blocks in run_vm.

Added BUGS-IN-GAMES: This seems to be necessary by now. Christoph,
could you change the FAQ to refer people to this file please? A
certain SQ3 bug may have to be added to it (I don't know the
specifics, so I didn't do it). Comments on the file itself are of
course welcome too.

(a) prevents Codename: Iceman from slowing to a crawl while on the
submarine.

Lars