Well, it's mostly a backout of my last commit. Got rid of the whole sdl_gfxprimitives library and restored our old linedraw code.
Added BOUNDS CHECKING to the linedraw scaling code; we were generating pixel values
outside the drawable surface if scaling was turned on.

So, please test.  Should fix bugs # 218 and #221.  And un-break all of the lines
that my last commit broke.

As an aside, giving scaling parameters on the command line with the sdl driver causes a coredump on my
box, I'm looking into it now..

 - Pizza