Lot's changed.
Thanks to Ravi's latest specs, the adlib instruments are now loaded
properly and don't make your ears bleed.

It now pumps out something resembling music, but with a few problems:

1) timing is a little shot.
2) operator/voice allocation

(1) I suspect is due to OSSseq not being all that great for realtime
    events.  I'm going to give alsa's sequencer a spin and see how that
    compares.
(2) We need to open up the channel priority to the midi_device layer.
    I'll probably tackle this later today.

 - Pizza