* Added BREAKPOINT() for __DECC and __alpha__ * Fixed heap corruption caused by size zero localvar blocks (a) * While cleaning up, the dynview list objects are verified with is_object() first (b)
(a): Apparently, our heap implementation doesn't support size zero allocations.
It now emits a warning if this is attempted.
(b): Old KQ4 appears to dispose some objects in inappropriate places; this
caused heap corruption.

Both (a) and (b) fix sources for heap corruption in KQ4. (b) might also fix the
problem in LSL2 when Larry is washed on the beach.
I'm not sure if this qualifies as a workaround or if it's what Sierra SCI does,
though.

-- Christoph