* Added global sci_open() and sci_fopen() functions for case-insensitive portable file opening * Made the resource loader use these after trying to open the lower-case versions
This means FreeSCI will always try resource.xxx first, then take a seemingly
random (as determined by readdir(3)) match if that doesn't work.

Regarding the chaseer problem with the witches: This feels a bit like the
problem in the CB1 intro, to be honest. I'm not sure what exactly there's
happening, but the mover thinks it's reached its target even though DoAvoider
notified it that it can't continue walking there (causing the mover to
finish by collission in the process of testing).

Personal note: I'll be away for a few days, possibly until thursday
(inclusive). I won't be having net access during that time, so any e-mails
I get won't receive an answer until about friday. Sorry for the inconvenience!

-- Christoph