* DoAvoider() no longer initializes b_movCnt (a) * DoBresen() no longer sets the mover's target coordinates to the client's coordinates if failing on collision (WHAT the hell was I thinking?) (b) * Some Win32 fixes for the resource manager (issues reported by Matt on IRC)
(a) fixes the chaser in HQ.
(b) closes (drum roll) #239 and #81.

Note on recent fixes: LSL2 and (now) KQ4 should be- theoretically- solveable
again.
(b) also fixes the problem with ego being stuck in the closet on the submarine
in Iceman, but a new problem (#242) is exhibited shortly after: The game is
stuck in an endless loop while Cdr Westland stands in front of the door.

HELP REQUIRED: What is supposed to happen at this point? Is the game just
supposed to give control back to the player, or is some sort of animation
(door opening, alert going off, 16 ton weight dropping down on ego)
supposed to happen first?

-- Christoph