* Split up sci.sgml into sci.sgml, freesci.sgml (FreeSCI-specific stuff) and sci-kernel.sgml (kernel functions) * Added short explanations of what the individual header files do to freesci.sgml, for almost all headers (except for old_objects.h and the sound headers) * Added an xslt stylesheet that processes kernel functions with a much simpler description into the current docbook style we use * Created sci-kernel-src.xml as an example (only contains Load() atm)
Eventually, I'd like to generate sci-kernel.sgml from sci-kernel-src.xml
and sci-kernel.xslt, because the structure in sci-kernel-src is easier
to use and also easier to transform into other languages (such as plain
text or html) than the docbook version we have now.
The xslt file still needs support for program listings, lists and tables,
though (and perhaps for a few other features as well).

The documentation updates have been propagated to the FreeSCI web page.


-- Christoph