* Experimental change: Added two alternative palettes
A question by Claudio lead me to this idea- AGI has two palettes, one for
Amiga and one for EGA, so he wondered whether SCI might have the same.
According to uae it doesn't, but that doesn't mean that we shouldn't try
how SCI would look with it...

To change the palette, enter 'set sci0_palette = 1' on the command console.
Palette 0 is standard EGA, 1 is Amiga-ish AGI, and 2 is a grayscale
palette using the the GIMP's luminance function. All other values will
have a reasonable chance of affecting your freesci processes negatively.


It's a quick hack ATM; please have a look and tell me if it's worth keeping
this (and cleaning it up). IMHO, KQ4 seems to benefit from it (not surprising,
considering its AGI origin), whereas the brighter, less saturated blue and
red values look a bit out of place in SQ3 and, to a lesser extent, QfG1.

-- Christoph