+ Added version check for priority OB1 problem in early SCI (a) + Swapped SOUNDSERVER_PIPED semantics in sound.c (b) + CanBeHere() now performs is_object() on cliplist objects (c) + Updated sfx_save struct for one 'short' variable (d) + Updated TODO: No sfx, but patch file support
Re (a): Fixes the new LSL3 priority bug without breaking the PQ2 doors (as
far as I can tell)
Re (b): I assume that's what it was supposed to do, it works here now.
Re (c): See BUGS-IN-GAMES
Re (d): Please be careful with changing types of existing variables if
references to them are passed as function parameters (there's a reason why
I rarely use anything other than 'int' for integers, and while part of it
is laziness and some other part is that it doesn't really matter in practice
if I use 'short' since it has to be aligned and stripped etc (at least on
RISC), it takes care of the problem of tracking variable types when passing
references of them... OK, so it _is_ laziness...)!


Just noticed nano is now my default editor. Funny.

-- Christoph