* Added Robey Pointer to THANKS and the 'about' menu for hosting the BTS (a) * Re-added support for resource patch files * song.c: Re-added the #ifdef SATISFY_PURIFY (b)
(a) He maintains lag.net, which our bug tracker is now being hosted on
(yes, it's back on-line!)

(b): The idea is that this might trigger warnings on some tools when
reading values that were not initialized properly. If there's a good reason
for doing the memset() always, it'd be OK to remove it again, of course.

Revised TODO for 0.3.3 is done, all that remains is bug-fixing and testing.

(Actually, we wouldn't really need a release- all it'd be good for would be
to gather bug reports, and we have enought of these already ;-)


-- Christoph