* Always set the savegame directory correctly during startup (i.e. _after_ initializing directories) (a) * When restoring, copy version, max_version, min_version, and the savegame/ working directories from the currently set values (b) * UNIX: Upon SIGTSTP/SIGCONT, pass the signal on to the sound server (c) * Minor fix in tools.c (d)
(a) and (b) solve a problem that made it impossible to restore or save again
after restoring a Win32 savegame under UNIX (or that would set a very weird
but correct working directory, after the (d) fix).

(b) should fix #289 and #290, please test!

(c) makes sure all of FreeSCI is suspended when pressing C-z on the shell,
fixing an (IIRC unfiled) bug reported by Claudio a while ago.


-- Christoph