* Renamed del -> delete in the selector list (a) * Reverted end-of-song handling back to the way it was two months ago (b)
(a) #ifndef __cplusplus; please report any problems this might cause!
(b) To summarize, the song status and position are set explicitly and
'newsong' is derived.

Could somebody test this on a system with sound support, please? I'd also be
interested in learning if this affects the LSL3 phone booth problem (not sure
if this generates a segfault on non-Win32).

Unless there are any problems with this, I assume it should be put into
event_ss.c as well, but please test the following:

SQ3: Intro, jumping out of the pod
QfG1: Character selection screen
LSL3: Phone booth

Please contact me even if it's not broken- TIA for any feedback!


-- Christoph