* Use text flags for calculating text widget sizes (a)
(a) should fix the KQ4 titlebar problem. This is a particularly interesting
problem, BTW:
KQ4 uses '\n' for the single "IV" character in its titlebar. Normally, this
character is interpretted as a newline; not so when printing to the title
bar, which draws the appropriate character, as specified in the font.

However, the text widget representing this text used to calculate the font
size without distinguishing between the title bar or any normal port (while
drawing correctly!), so it believed the text was much smaller than it
actually was, and therefore didn't bother to redraw it when it should have.

That's the secret behind the partially updating title bar :-)


-- Christoph