* Fixed some crap (a) in pic drawing code (b)
(a) Can't be described as anything else.
(b) Pattern plotting into unscaled buffers, to be precise. Fixes the
PQ2 drawpic segfault and the unfilled area in CB1 (although I didn't intend
it to do the latter).


(I'm ALMOST sure that this fix is neccessary, right and won't break
anything that didn't deserve to be broken in the first place, but I
reserve the right to roll this one back during the freeze if it turns into
a show stopper!)

-- Christoph