2002-02-19  Alexander R Angas <wgd@internode.on.net>
	* src/config.c:
	* src/win32/freesci.def: Fixes to Win32 build.

	* src/sound/buff_ss.c: (new)
	* src/sound/buff_ss_win32.c: (new)
	* src/sound/midiout_win32mci_stream.c: (new)
	* src/include/midiout.h:
	* src/include/sound.h:
	* src/sound/event_ss.c:
	* src/sound/midiout.c:
	* src/sound/sound.c:
	* src/sound/soundserver.c:
	* src/win32/fsci_dll.dsp: Added prototype Win32 MCI stream driver and
	sound server to work with it.

This code doesn't work and some of it will be removed soon in light of recent discussion.
While decisions over the next sound overhaul are being thought about and made,
I hope to have this MCI stream driver at least playing one song as the code for
the driver itself will be needed after the overhaul.

 -- Alex.