We now have automagic interpreter-version detection.
This is accomplished through a lookup table, based on the size of the
sciv.exe/sierra.exe file.  So far, it's unique.

So far I have thirteen different interpreter versions in the table for
SCI0 games, I'm sure a few more can be added later on.

(We could alternatively do a checksum/size on the resource.000/001 file
 instead.  Comments?)

This is currently only overridable on the command line.

 - Pizza