* xlib driver now uses XNextEvent() to clear the X11 event queue, when appropriate
I currently believe that this will fix the slowdown problems when playing
for extended periods of time. Please test!

n.b.: Note that this should include a general speed-up when running with
the xlib driver.

Now all we'd need would be accellerated X11 calls for pixmap scaling, as
seen in Episode IV: Owl-stretching time.

-- Christoph