ALSA09 pcmout driver now works.
Trashed the async callback model that screwed up the soundserver due to
SIGIO collisions.

Instead, we use pthreads and use a background thread to do all the work.
Works pretty well, though some experimentation will be needed to find
the optimal number of periods/buffers.

We also may need to update the autoconf scripts to reflect the pthreads
dependency for this device.

 - Pizza