PCM subsystem now has stereo/mono as a run-time argument.  When in mono mode, the adlibemu uses half the CPU power.  go figure.  :)
Note that I didn't update config.c, as my lex does it differently than
the one that was checked into CVS last time.  So I'll let someone else
take care of that.

This is almost everything on my plate except for the new soundserver.
that'll come in a little while.  :)

 - Pizza