(a) Implemented sound stubs for SCI01 and SCI1. (b) Removed all calls to CHECK_THIS_KERNEL_FUNCTION.
(a) marks the beginning of the adaptation of the sound subsystem to
later SCI versions. It also includes a temporary hack which allows the
QfG2 intro to continue past the point where it used to lock up. This,
of course, should be removed when the stubs are replaced with real
code. The stubs include some dummy code which is meant to guide the
implementation.

(b) is just a convenience for developers who want to monitor the
activity of one or a few kernel calls.

Lars