Experimental fix for the high cpu utilization problem with the threaded soundserver.
On win32, we perform a sched_yield() when we enter our "delay" loop.  On
everything else, we usleep(0) if we are to sleep for greater than 10ms,
otherwise we just busyloop.    CPU usage on my laptop is way down and
note timing seems to be sane still.

We need to experiment with the magic number a bit.  10ms isn't quite
arbitrary, but there may be a lower number that is more optimal.

 - Pizza