* Misc. bugfixes in the gfx console, wrt scrolling * viewinfo now dumps cels along with view meta-information when on the   graphical console * disasm can now be passed flags to request the hexadecimal representation   of the numerical values of the bytecode it disassembles to be set, and   no longer displays the Word/Byte flag information by default (although   there is another flag that allows this still to be print).
For reference: You need to configure --with-console for the gfx console
to be built, because of the performance and memory overhead incurred by
it.
The graphical console is used by default on any system that does not have
the new GFX_CAPABILITY_WINDOWED capability flag set. Right now, only
the xlib driver has this set because it's the only driver of which I know
that it does not support fullscreen for certain.
SDL, DirectX and GGI are the ones that might want to change this, according
to whether they are run as full-screen.

-- Christoph