* Fixed script instantiation * Fixed some stuff in the VM * Misc. fixes in game.c and vm.c to take advantage of segmentation * Added segtable, seginfo, vo debugger calls * Extended debugger to accept type 'a' arguments (addresses) * Cleaned up console record management * Separated manual pages into sections 1 (commands), 2 (variables),   and 3 (misc. documents) * Added support for looking up classes by name in the debugger   (for commands taking 'address' type parameters, i.e. only vo ATM)
Stuff seems to be working a bit more reasonably now. If you care about
name lookup, have a look at 'addresses.3' in the manual pages.
Note that addresses are in hexadecimal, whereas script numbers are
(usually) in decimal.

  Next thing to do is to introduce stack, list, and list node segments
and get the first couple of kernel functions working again (while
debugging).

-- Christoph