* Fixed segfault in function selector lookup
It's now possible to enter the QfG2 character selection screen, although this
triggers a segfault elsewhere (graphics code). Disabling that has that
screen mostly not working, because formatting prints seem to be broken for
some reason.

I guess qfg2 resets afterwards, because continuing anyway causes a weird
segfault which is hard to explain otherwise...

But, as intended, it does not run out of memory.

-- Christoph