a) Fixed argument base in AvoidPath and relaxed the kernel signature. b) Fixed signedness issue in the class opcode.
a) fixes movement in the inn, and AvoidPath() now displays the
warnings it's supposed to.

b) There was a similar bug in super. I hope this is the last one of
them.

The kernel signatures are starting to get on my nerves! Many kernel
functions have alternate syntaxes which cause the signature check to
fail. The easiest way to fix this is to relax the kernel signatures,
but that kiod of defeats their purpose.

Lars