(a) fixed a few Valgrind warnings (b) Fixed segfault when restarting KQ1 (by porting DrawStatus) (c) Made the disassembly handle &rest arguments better
(b) is a sign of a bug in the kernel emulation layer. I am not going
to hunt it down and fix it unless someone really urges me to; the
long-term goal is to get rid of this code altogether. I guess we can
add this to the TODO for glutton 0.6.1.

Lars