2003-03-19  Lars Skovlund  <lskovlun@image.dk>
	* src/engine/kscripts.c (kClone): The script of the parent object
	gets locked on each clone. (a)
	* src/engine/game.c (script_init_engine): No longer add stray
	classes that are not in the class table. (b)

(a) works around the locking problem in QfG2 (after the last update, a
superclass was being disposed with the subclass still present) by
assuring that scripts containing classes that are typically cloned
will have a high lock count. This isn't the consistent use of
locking/unlocking that I've advocated, but it does the job for now.

(b) works around a bug in SQ4 floppy where an older version (read:
broken) of a script would take precedence over the correct
script.

Lars