* alsa-midi driver timing fixes * When starting a new song, instruct the output driver to reset its timer   (if supported) (a) * When programming an MT-32, make sure it has enough time to process all   commands (b)
(a) fixes issues of song timing being off, making timing much more in-
dependent of the underlying timing mechanism (particularly important with
our current default method).

(b) If an output driver does not support flushing, we now estimate the time
the MT-32 needs to complete processing its instructions (and ensure that
the individual sysexes we send are delayed appropriately); if the output
timer is being reset (in preparation for a song starting), we defer
resetting it (and doing anything else) until the MT-32 has finished being
programmed, as far as we can tell from our estimates.

This should fix the "fast" MT-32 programming we got in Glutton without
breaking apart for particularly long sysex sequences.

-- Christoph