* Added PCM device, PCM data feed (input source) and PCM mixer interfaces * Added software PCM mixer implementation (using linear interpolation)
I'd appreciate any feedback regarding the interface design.
The basic idea wrt pcm sequencers will be to have an input feed act as a
mediator, receiving MIDI events, polling sequencers for the MIDI delay
tiems and storing the results, and passing on actual MIDI commands to them
in order to allow changes of behaviour.

-- Christoph