* Retired old software sequencer stuff * Implemented different software sequencer stuff based on the notion of   sequencers being polled by the _mixer_ * Ported PC speaker sequencer to new layer
The new layer gets rid of the "intermediate" audio buffers that used to be
required because sequencers couldn't write directly to buffers used by the
mixer. Being polled separately, they had to write ahead some unknown number
of samples. Now, we only write as many samples as the mixer asks for,
getting rid of a lot of complexity.

-- Christoph