* Added more strict checking to SCI1 vocabulary generation (a) * Avoid NULL pointer dereferentiation when PCM subsystem fails to initialise * Added 'additional_frames' write-ahead for the soft PCM mixer (b) * Added a number of SCI games reported by Jonathan Gray to our SCI games   list * Temporarily added some mem testing to the SCI1 iterator
(a) probably fixes the JITFL out-of-bounds read.
(b) Once configurable, this allows additional bytes to be written ahead in
mixer computations. It balances minor differences between the number of
frames estimated by the mixer and the number of frames required by the
PCM player (certainly sounds better on my Linux iBook now).

-- Christoph