Fix port handling in DisposeWindow() so it always chooses the window manager port to be active afterwards. This fixes some situations in which DrawStatus() was called from inside Animate(), found in SQ3 and other games.
Oh, and Roger falls off the rail again. We used to fix this by not
taking the y-coordinate base setter adjustment into account, but that
is obviously wrong. I removed this "fix" some weeks ago, and I'm not
going to put it back in unless someone thinks it's really important to
have SQ3 working.

Lars