* Added Ismail's patch:   - PocketPC portability fixes   - On-screen keyboard for PocketPC/SDL   - Non-integral screen scaling for PocketPC/SDL   - Game selection screen * Modulo the following changes:   - Game selection screen is now globally accessible   - Game slection screen code in src/menu   - Game selection screen fonts in src/gfx   - File creation relegated to new sci_create_file() function
I have no means to test whether this actually works. The patch was fairly
big, and I modified several things while applying it-- please test and let
me know if anything is missing!

Also: DC and PocketPC now both have different copies of the keyboard code.
It would be better for the platforms needing it if both code and data could
be shared...

-- Christoph