* Added a simple FSM to detect version number strings in binary executables * Changed our "CRC" code to a conventional hashing function * Added a handful of games and hash codes to games.h * Updated TODO list to reflect the changes (1 down, 2 to go)
  The hashing code is the string hashing function everyone seems to use in
their standard library (foldl (\hash. \ch. hash * c1 + ch) c2); c2 here is
some constant derived from the RESOURCE file found, and c1 is 19 (which is
what the magical O'Caml builtin hashing function uses on strings; IIRC all
that matters is that it's a reasonably small prime number).

  You may now get warnings regarding noninterpreted hash codes for games
for which autodetection used to work; apologies for this.

-- Christoph