Differentiate between unknown and unmapped/NOP functions. Relegating unmapped functions to k_Unknown is not the right thing to do ATM, because we are using the vocab resource (999.voc) and trust it to be accurate*. IMO, we'll eventually have to make a decision whether to rely completely on the kernel vocabs or to ditch them completely. The hybrid scheme we currently use is confusing and difficult to maintain.
Matt: For KQ5/CD, a special patch file (a hand-generated 999.voc) is
needed to run. There's not much to see either; the game hangs at the
Sierra logo, waiting for an SCI1.1-style digital sound cue (because
that part hasn't been coded yet). The 'restart' trick works, but there
are major problems controlling ego. I suspect a bug in either
gfxop_set_pointer_position or in our cursor hotspot logic.

OTOH, this was a nice way to test the view cursor support ahead of time!

Lars

* which it isn't always; this is the root cause of our problems with
"unknown" kernel functions. A better term for these might be
"unlisted", but there are also cases (such as KQ5/CD, at which this
patch is targeted) in which some entries are just plain wrong.