Implement support for the array handling style used by the SCI MCyc class. This is most obvious in the SQ4 floppy intro, but use of the class has been spotted elsewhere also.
To simplify the above, I have standardized on a single reg_t
representation. The "new" representation used to be an #ifdef used on
some platforms. I can't see why the old default should be preferable,
and it seems to work with a variety of games. If anyone has
objections, don't hesitate to say so.

Lars